+++ miner/proposal_builder.go
@@ -373,6 +373,17 @@ func (pb *ProposalBuilder) handleLayer(ctx context.Context, layerID types.LayerI
 	if !pb.syncer.IsSynced(ctx) {
 		return errNotSynced
 	}
+
+	// make sure the miner is eligible first
+	nonce, err := pb.nonceFetcher.VRFNonce(pb.signer.NodeID(), layerID.GetEpoch())
+	if err != nil {
+		if errors.Is(err, sql.ErrNotFound) {
+			pb.logger.WithContext(ctx).With().Info("miner has no valid vrf nonce, not building proposal", layerID)
+			return nil
+		}
+		return err
+	}
+
 	if beacon, err = pb.beaconProvider.GetBeacon(epoch); err != nil {
 		return errNoBeacon
 	}
@@ -386,14 +397,6 @@ func (pb *ProposalBuilder) handleLayer(ctx context.Context, layerID types.LayerI
 		return errDuplicateLayer
 	}
 
-	nonce, err := pb.nonceFetcher.VRFNonce(pb.signer.NodeID(), layerID.GetEpoch())
-	if err != nil {
-		if errors.Is(err, sql.ErrNotFound) {
-			pb.logger.WithContext(ctx).With().Info("miner has no valid vrf nonce, not building proposal", layerID)
-			return nil
-		}
-		return err
-	}
 	epochEligibility, err := pb.proposalOracle.ProposalEligibility(layerID, beacon, nonce)
 	if err != nil {
 		if errors.Is(err, errMinerHasNoATXInPreviousEpoch) {
@@ -470,8 +473,12 @@ func (pb *ProposalBuilder) createProposalLoop(ctx context.Context) {
 			}
 			next = current.Add(1)
 			lyrCtx := log.WithNewSessionID(ctx)
-			if err := pb.handleLayer(lyrCtx, current); err != nil && !errors.Is(err, errGenesis) {
-				pb.logger.WithContext(lyrCtx).With().Warning("failed to build proposal", current, log.Err(err))
+			if err := pb.handleLayer(lyrCtx, current); err != nil {
+				switch {
+				case errors.Is(err, errGenesis), errors.Is(err, errNotSynced):
+				default:
+					pb.logger.WithContext(lyrCtx).With().Warning("failed to build proposal", current, log.Err(err))
+				}
 			}
 		}
 	}